FIX: OYO 관리 및 생성 오류 수정
